Update limits across app
